auth-backend: move OAuthEnvironmentHandler to auth-node
